Make loose parser work with minor interface changes introduced by modularization
